Moved dataProcessing.tsx and semVer.tsx into src and changed file type to .ts
These files do not define react components, as such they should not be in the components folder and not use .tsx
